Added hover, improved layout

+++ frontend/src/components/AccessTable.jsx
@@ -226,41 +226,49 @@ const AccessTable = () => {
     window.location.reload();
   };
 
+  const renderActions = (user) => {
+    if (user.role !== 'Admin') {
+      return (
+        <>
+          <button className='bg-gray-100 text-blue-700 mr-2 hover:scale-105 px-4 py-2' onClick={() => toggleEditCard(user)}>✏️</button>
+          <button
+            className='bg-gray-100 text-red-700 hover:scale-105 px-4 py-2'
+            onClick={() => deleteUser(user.id)}
+          >🗑️</button>
+        </>
+      );
+    } else {
+      return (
+        <p className='mr-2 p-[20px]'> </p>
+      );
+    }
+  }
+
   return (
-    <div className='mt-12 flex flex-col'>
+    <div className='mt-8 flex flex-col'>
       <div className='flex mt-8 ml-2 justify-between w-[930px]'>
         <p className='font-bold text-3xl text-blue-700 mb-8'>Access</p>
-        <button className='bg-blue-700 text-white w-[100px] p-0 rounded h-[30px] text-sm' onClick={toggleAddCard}>Add User</button>
+        <button className='bg-blue-700 text-white w-[110px] p-0 rounded h-[34px] text-base hover:scale-105' onClick={toggleAddCard}>Add User</button>
       </div>
-      <table className='w-[1110px] table-auto border-collapse'>
+      <table className='w-[1110px] table-auto border-collapse ml-4'>
         <thead>
           <tr>
-            <th className='text-left'>Name</th>
-            <th className='text-left'>Email</th>
-            <th className='text-left'>Role</th>
-            <th className='text-left'>Actions</th>
+            <th className='text-left w-[140px] pb-[10px]'>Name</th>
+            <th className='text-left w-[190px] pb-[10px]'>Email</th>
+            <th className='text-left w-[130px] pb-[10px]'>Role</th>
+            <th className='text-left w-[170px] pb-[10px]'>Actions</th>
           </tr>
         </thead>
         <tbody>
           {users.map((user, index) => (
             <tr key={index}>
-              <td className='py-2'>{user.name}</td>
-              <td className='py-2'>{user.email}</td>
-              <td className='py-2'>
+              <td className='py-2 w-[140px]'>{user.name}</td>
+              <td className='py-2 w-[190px]'>{user.email}</td>
+              <td className='py-2 w-[130px]'>
                 <span className={roleStyles[user.role]}>{user.role}</span>
               </td>
-              <td className='py-2'>
-                {user.role !== 'Admin' && (
-                  <>
-                    <button className='bg-gray-100 text-blue-700 mr-2' onClick={() => toggleEditCard(user)}>✏️</button>
-                    <button
-                      className='bg-gray-100 text-red-700'
-                      onClick={() => deleteUser(user.id)}
-                    >
-                      🗑️
-                    </button>
-                  </>
-                )}
+              <td className='py-2 w-[170px]'>
+                {renderActions(user)}
               </td>
             </tr>
           ))}
@@ -277,7 +285,7 @@ const AccessTable = () => {
         />
       )}
     </div>
-  );
+  );  
 };
 
 export default AccessTable;
